refactor: duckdbAggregator.tsをaggregator.tsにリネームし型定義を統合
aggregate.ts(型定義のみ)をaggregator.tsに統合して削除。
